pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "edgent-smith"
version = "0.1.0"
description = "Agentic system optimized for edge models"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "edgent-smith contributors" }]
dependencies = [
    "pydantic-ai>=0.0.14",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",
    "httpx>=0.27",
    "anyio>=4.4",
    "structlog>=24.1",
    "prometheus-client>=0.20",
    "python-ulid>=2.2",
    "tenacity>=8.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "httpx>=0.27",
    "mypy>=1.10",
    "ruff>=0.4",
    "types-requests>=2.31",
]

[project.scripts]
edgent-smith = "edgent_smith.main:run"

[tool.hatch.build.targets.wheel]
packages = ["src/edgent_smith"]

[tool.ruff]
src = ["src"]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
exclude = ["tests/", "eval/", "experiments/"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--tb=short -q"

[tool.coverage.run]
source = ["src/edgent_smith"]
